Move Algo file into its own component
